fix warning about int vs. float in moveable.cpp
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@2082 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/moveable.cpp
@@ -222,7 +222,7 @@ void idMoveable::Save( idSaveGame *savefile ) const {
 	savefile->WriteBool( canDamage );
 	savefile->WriteInt( nextDamageTime );
 	savefile->WriteInt( nextSoundTime );
-	savefile->WriteInt( initialSpline != NULL ? initialSpline->GetTime( 0 ) : -1 );
+	savefile->WriteInt( initialSpline != NULL ? (int)initialSpline->GetTime( 0 ) : -1 );
 	savefile->WriteVec3( initialSplineDir );
 
 	savefile->WriteStaticObject( physicsObj );
